+++ app/src/main/java/com/sergeyfitis/recyclerviewdemo/easy_view_type/adapter/models/BigPictureViewItem.java
@@ -23,7 +23,7 @@ public class BigPictureViewItem implements ViewItem {
     }
 
     @Override
-    public int distinctId() {
+    public long distinctId() {
         return item.hashCode();
     }
