+++ app/src/main/java/com/sergeyfitis/recyclerviewdemo/easy_view_type/adapter/models/CommonViewItem.java
@@ -25,7 +25,7 @@ public class CommonViewItem implements ViewItem {
     }
 
     @Override
-    public int distinctId() {
+    public long distinctId() {
         return commonItem.hashCode();
     }
