+++ app/src/main/java/com/sergeyfitis/recyclerviewdemo/easy_view_type/adapter/models/HeaderViewItem.java
@@ -24,7 +24,7 @@ public class HeaderViewItem implements ViewItem {
     }
 
     @Override
-    public int distinctId() {
+    public long distinctId() {
         return headerText.hashCode();
     }
